Initial creation of selection-module to hold layer copies
